+++ Library/OcSmbiosLib/SmbiosInternal.c
@@ -461,7 +461,7 @@ SmbiosGetStructureLength (
   Walker = SmbiosTable.Raw;
   while (SmbiosTableSize >= SMBIOS_STRUCTURE_TERMINATOR_SIZE) {
     if (Walker[0] == 0 && Walker[1] == 0) {
-      return Length + (Walker - SmbiosTable.Raw) + SMBIOS_STRUCTURE_TERMINATOR_SIZE;
+      return Length + (UINT32)(Walker - SmbiosTable.Raw) + SMBIOS_STRUCTURE_TERMINATOR_SIZE;
     }
     Walker++;
     SmbiosTableSize--;
